assign/Assistant_api/personal_trainer/requirements.txt
openai>=1.0.0
python-dotenv
httpx==0.27.2